feat(#271): blind-reviewer-reachable trace review surface (UI)
Wires the domain-expert (blind) reviewer to the M31.4 trace review, which was
unreachable — the trace pages lived under ProjectLayout, which redirects blind
evaluators to /eval. Mirrors the prompt-output blind surface (SessionDeck):
routes outside ProjectLayout, no org/project context, read from the query
payload.

- TraceReviewBody / TraceMatchupBody: shared, context-free review bodies
  extracted from TraceViewer/TraceMatchup (VerdictControl moved in). Owner routes
  and blind routes now render the exact same surface; the owner wrappers keep
  useProject, the blind pages don't.
- New blind pages under /eval (EvalLayout, no project context): TraceReviewList
  (discovery via listReviewableTraces), TraceReview (:agentTraceId, title
  "Evaluation — {project}" per blind-eval Rule 3), TraceMatchupReview.
- Routes: /eval/traces, /eval/traces/:agentTraceId, /eval/matchups/:matchupId.
- InvitesInbox: a "N trajectories to review →" card on the /eval landing so the
  surface is discoverable (renders only when non-empty).

Blind-safety: for a blind reviewer the header shows "Trajectory" (no harness/
model/product), the bias-reduction framing card, no data-* provenance, blue/
purple only.

Playwright CT: new blind review-page render smoke (no provenance, bias framing,
verdict, steps) + reviewable-list render, alongside the existing step specs —
4 CT specs green. npm run build clean.
