+++ src/routes/traces/TraceMatchup.tsx
@@ -1,32 +1,13 @@
 /**
- * #267 (M31.4): step-level pairwise preference. Two blind-labelled trajectories
- * side by side, sharing the same task up to a divergence point; the reviewer
- * picks whose NEXT move is better. Both sides page steps through the same
- * blind-projected listSteps — no provenance reaches the DOM.
+ * #267 (M31.4): step-level pairwise preference for owners and editors, nested
+ * under ProjectLayout. The matchup surface lives in the shared TraceMatchupBody
+ * (#271) so blind reviewers reach an identical view via /eval/matchups/:matchupId.
+ * This wrapper only supplies the project-scoped back link.
  */
-import { useEffect, useState } from "react";
-import { useQuery, useMutation } from "convex/react";
-import { Link, useParams } from "react-router-dom";
-import { api } from "../../../convex/_generated/api";
+import { useParams } from "react-router-dom";
 import type { Id } from "../../../convex/_generated/dataModel";
 import { useProject } from "@/contexts/ProjectContext";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-import { Skeleton } from "@/components/ui/skeleton";
-import { buttonVariants } from "@/components/ui/button";
-import { cn } from "@/lib/utils";
-import { friendlyError } from "@/lib/errors";
-import { StepList, COMMENT_TAGS, type CommentTag } from "./traceSteps";
-import { ArrowLeft } from "lucide-react";
-
-const WINNERS = [
-  { value: "left", label: "◀ Left better" },
-  { value: "right", label: "Right better ▶" },
-  { value: "tie", label: "Tie" },
-  { value: "skip", label: "Skip" },
-] as const;
-
-type Winner = (typeof WINNERS)[number]["value"];
+import { TraceMatchupBody } from "./TraceMatchupBody";
 
 export function TraceMatchup() {
   const { projectId } = useProject();
@@ -36,179 +17,13 @@ export function TraceMatchup() {
   }>();
   const id = matchupId as Id<"agentTraceMatchups">;
 
-  const matchup = useQuery(api.agentTraceReview.getMatchup, { matchupId: id });
-  const decide = useMutation(api.agentTraceReview.decideMatchup);
-
-  const [tags, setTags] = useState<CommentTag[]>([]);
-  const [tagsInit, setTagsInit] = useState(false);
-  const [busy, setBusy] = useState(false);
-  const [error, setError] = useState("");
-
-  useEffect(() => {
-    if (!tagsInit && matchup) {
-      setTags(matchup.reasonTags as CommentTag[]);
-      setTagsInit(true);
-    }
-  }, [matchup, tagsInit]);
-
   const projectBase = `/orgs/${orgSlug}/projects/${projectId}`;
 
-  if (matchup === undefined) {
-    return (
-      <div className="mx-auto max-w-5xl space-y-4 p-6">
-        <Skeleton className="h-8 w-64" />
-        <div className="grid gap-4 md:grid-cols-2">
-          <Skeleton className="h-96 w-full" />
-          <Skeleton className="h-96 w-full" />
-        </div>
-      </div>
-    );
-  }
-
-  if (matchup === null) {
-    return (
-      <div className="mx-auto max-w-3xl space-y-3 p-6">
-        <p className="text-sm">
-          This matchup isn’t available. It may have been removed, or you may not
-          have access to it.
-        </p>
-        <Link
-          to={`${projectBase}/traces`}
-          className={buttonVariants({ size: "sm", variant: "outline" })}
-        >
-          <ArrowLeft aria-hidden="true" className="mr-1.5 h-3.5 w-3.5" />
-          Back to trajectories
-        </Link>
-      </div>
-    );
-  }
-
-  function toggleTag(tag: CommentTag) {
-    setTags((prev) =>
-      prev.includes(tag) ? prev.filter((t) => t !== tag) : [...prev, tag],
-    );
-  }
-
-  async function submit(winner: Winner) {
-    setBusy(true);
-    setError("");
-    try {
-      await decide({ matchupId: id, winner, reasonTags: tags });
-    } catch (err) {
-      setError(friendlyError(err, "Couldn’t record your pick. Try again."));
-    } finally {
-      setBusy(false);
-    }
-  }
-
   return (
-    <div className="mx-auto max-w-5xl p-6">
-      <Link
-        to={`${projectBase}/traces`}
-        className="inline-flex items-center gap-1 text-sm text-muted-foreground hover:text-foreground"
-      >
-        <ArrowLeft aria-hidden="true" className="h-3 w-3" />
-        Trajectories
-      </Link>
-
-      <header className="mt-3">
-        <h1 className="text-2xl font-bold">Which next move is better?</h1>
-        <p className="mt-1 text-sm text-muted-foreground">
-          Given the same task up to this point, which agent’s next move is
-          better?
-        </p>
-      </header>
-
-      <div className="mt-6 grid gap-4 md:grid-cols-2">
-        <MatchupColumn
-          heading={matchup.leftBlindLabel}
-          agentTraceId={matchup.leftTraceId}
-          divergenceStepIndex={matchup.divergenceStepIndex}
-        />
-        <MatchupColumn
-          heading={matchup.rightBlindLabel}
-          agentTraceId={matchup.rightTraceId}
-          divergenceStepIndex={matchup.divergenceStepIndex}
-        />
-      </div>
-
-      <Card className="mt-6">
-        <CardHeader>
-          <CardTitle className="text-base">Your pick</CardTitle>
-        </CardHeader>
-        <CardContent className="space-y-3">
-          <div className="flex flex-wrap gap-2">
-            {WINNERS.map((w) => {
-              const selected = matchup.winner === w.value;
-              return (
-                <Button
-                  key={w.value}
-                  type="button"
-                  variant={selected ? "default" : "outline"}
-                  size="sm"
-                  onClick={() => void submit(w.value)}
-                  disabled={busy}
-                  className={cn(selected && "ring-2 ring-primary/40")}
-                >
-                  {w.label}
-                </Button>
-              );
-            })}
-          </div>
-
-          <div>
-            <p className="mb-1.5 text-xs text-muted-foreground">
-              Reasons (optional)
-            </p>
-            <div className="flex flex-wrap gap-1.5">
-              {COMMENT_TAGS.map((tag) => {
-                const on = tags.includes(tag);
-                return (
-                  <button
-                    key={tag}
-                    type="button"
-                    onClick={() => toggleTag(tag)}
-                    className={cn(
-                      "rounded-full border px-2 py-0.5 text-xs transition-colors",
-                      on
-                        ? "border-primary bg-primary/10 text-primary"
-                        : "border-border text-muted-foreground hover:text-foreground",
-                    )}
-                  >
-                    {tag}
-                  </button>
-                );
-              })}
-            </div>
-          </div>
-
-          {error && (
-            <p className="text-sm text-destructive" role="alert">
-              {error}
-            </p>
-          )}
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
-
-function MatchupColumn({
-  heading,
-  agentTraceId,
-  divergenceStepIndex,
-}: {
-  heading: string;
-  agentTraceId: Id<"agentTraces">;
-  divergenceStepIndex: number;
-}) {
-  return (
-    <div className="min-w-0">
-      <h2 className="mb-3 text-sm font-semibold">{heading}</h2>
-      <StepList
-        agentTraceId={agentTraceId}
-        divergenceStepIndex={divergenceStepIndex}
-      />
-    </div>
+    <TraceMatchupBody
+      matchupId={id}
+      backTo={`${projectBase}/traces`}
+      backLabel="Trajectories"
+    />
   );
 }
