+++ src/routes/traces/__smoke__/convexReactMock.tsx
@@ -4,6 +4,8 @@
  * lazy-expand against deterministic fixture data — no backend, no auth. Only the
  * hooks the trace components import are stubbed.
  */
+import { getFunctionName } from "convex/server";
+
 export const FIXTURE_STEPS = [
   { stepIndex: 0, kind: "message", role: "assistant", hasBody: true },
   {
@@ -51,6 +53,59 @@ export function useMutation() {
   return noopMutation;
 }
 
-export function useQuery() {
+// Fixture rows for the blind reviewer's discovery list (listReviewableTraces).
+export const FIXTURE_REVIEWABLE = [
+  {
+    _id: "trace_review_1",
+    projectName: "Support Router",
+    status: "ready",
+    stepCount: 12,
+    createdAt: 2,
+  },
+  {
+    _id: "trace_review_2",
+    projectName: "Refund Agent",
+    status: "ready",
+    stepCount: 3,
+    createdAt: 1,
+  },
+];
+
+// A blind reviewer's getTrace payload — provenance stripped by the backend
+// (harness/model/product undefined), projectName + usage retained.
+export const FIXTURE_BLIND_TRACE = {
+  _id: "trace_review_1",
+  projectName: "Support Router",
+  traceId: undefined,
+  product: undefined,
+  module: undefined,
+  environment: undefined,
+  status: "ready",
+  stepCount: 4,
+  privacyClass: "internal",
+  model: undefined,
+  harnessName: undefined,
+  harnessVersion: undefined,
+  usage: { totalTokens: 900 },
+  hasFinalAnswer: true,
+};
+
+// Branch by function name so each query gets the right fixture; anything
+// unmapped stays undefined (loading), keeping specs independent.
+export function useQuery(query: unknown) {
+  try {
+    switch (getFunctionName(query as never)) {
+      case "agentTraces:listReviewableTraces":
+        return FIXTURE_REVIEWABLE;
+      case "agentTraces:getTrace":
+        return FIXTURE_BLIND_TRACE;
+      case "agentTraceReview:myVerdict":
+        return null;
+      case "agentTraceReview:listComments":
+        return [];
+    }
+  } catch {
+    // Not a resolvable function reference — treat as loading.
+  }
   return undefined;
 }
